17.02

+++ src/main/java/com/example/epamcourse/controller/command/impl/PickUpDocumentsCommand.java
@@ -1,22 +1,24 @@
 package com.example.epamcourse.controller.command.impl;
 
 import com.example.epamcourse.controller.command.*;
+import com.example.epamcourse.model.entity.Certificate;
+import com.example.epamcourse.model.entity.Faculty;
+import com.example.epamcourse.model.entity.Subject;
 import com.example.epamcourse.model.exception.CommandException;
 
 import com.example.epamcourse.model.exception.ServiceException;
-import com.example.epamcourse.model.service.ApplicantService;
-import com.example.epamcourse.model.service.BillService;
-import com.example.epamcourse.model.service.CertificateService;
-import com.example.epamcourse.model.service.SubjectService;
-import com.example.epamcourse.model.service.impl.ApplicantServiceImpl;
-import com.example.epamcourse.model.service.impl.BillServiceImpl;
-import com.example.epamcourse.model.service.impl.CertificateServiceImpl;
-import com.example.epamcourse.model.service.impl.SubjectServiceImpl;
+import com.example.epamcourse.model.service.*;
+import com.example.epamcourse.model.service.impl.*;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+
+import java.util.List;
+import java.util.Optional;
+
+import static com.example.epamcourse.controller.command.PagePath.MAIN_PAGE_APPLICANT_REDIRECT;
 
 public class PickUpDocumentsCommand implements Command {
     private static final Logger logger = LogManager.getLogger();
@@ -24,7 +26,7 @@ public class PickUpDocumentsCommand implements Command {
     @Override
     public Router execute(HttpServletRequest request) throws CommandException {
         Router router = new Router(PagePath.MAIN_PAGE_APPLICANT_REDIRECT);
-        ApplicantService applicantService = ApplicantServiceImpl.getInstance();
+        RecruitmentService recruitmentService = RecruitmentServiceImpl.getInstance();
         BillService billService = BillServiceImpl.getInstance();
         SubjectService subjectService = SubjectServiceImpl.getInstance();
         CertificateService certificateService = CertificateServiceImpl.getInstance();
@@ -32,14 +34,21 @@ public class PickUpDocumentsCommand implements Command {
         session.setAttribute(SessionAttribute.CURRENT_PAGE, PagePath.MAIN_PAGE_APPLICANT_REDIRECT);
         Long applicantId = (Long) session.getAttribute(SessionAttribute.APPLICANT_ID);
         try {
-            if (!billService.isBillArchive(applicantId)) {
-                billService.deleteBill(applicantId);
-                subjectService.deleteSubjects(applicantId);
-                certificateService.deleteCertificate(applicantId);
-                request.setAttribute(RequestAttribute.MESSAGE, LocaleMessageKey.PICK_UP_DOCUMENTS_SUCCESS);
+            if (recruitmentService.isRecruitmentActive()) {
+                if (!billService.isBillArchive(applicantId)) {
+                    billService.deleteBill(applicantId);
+                    subjectService.deleteSubjects(applicantId);
+                    certificateService.deleteCertificate(applicantId);
+                    request.setAttribute(RequestAttribute.MESSAGE, LocaleMessageKey.PICK_UP_DOCUMENTS_SUCCESS);
+                } else {
+                    router.setPage(MAIN_PAGE_APPLICANT_REDIRECT); // todo
+                    session.setAttribute(RequestAttribute.MESSAGE, LocaleMessageKey.ADD_REQUEST_ERROR);
+                }
             } else {
-                request.setAttribute(RequestAttribute.MESSAGE, LocaleMessageKey.PICK_UP_DOCUMENTS_ERROR);
+                router.setPage(MAIN_PAGE_APPLICANT_REDIRECT);
+                session.setAttribute(RequestAttribute.MESSAGE, LocaleMessageKey.RECRUITMENT_NOT_STARTED);
             }
+
         } catch (ServiceException e) {
             logger.log(Level.ERROR, "Pick up documents failed", e);
             throw new CommandException("Pick up documents failed", e);
